requirements.txt
streamlit
tabula-py
pandas
langchain
python-dotenv
openai
tabulate